+++ mock/scripts/linux/00-05_Duplicates.sh
@@ -15,4 +15,4 @@ if [ ! -d output ]; then
     mkdir output
 fi
 
-java -jar ../deploy/ELS.jar -C . -c debug -d debug -p libraries/publisher.json -T libraries/targets.json -F output/00-05_Duplicates.log --duplicates
+java -jar ../deploy/ELS.jar -C . -c debug -d debug -p libraries/publisher.json -F output/00-05_Duplicates.log --duplicates